requirements.txt
paddleocr
paddlepaddle
protobuf==3.20.0
opencv-python-headless
opencv-python
google-generativeai
Pillow
numpy
python-dotenv
fastapi
uvicorn[standard]
